backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.119.1
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic-settings>=2.10.1
pydantic_core==2.41.4
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
python-dotenv==1.0.0
imapclient==3.0.1
aiosmtplib==3.0.2
email-validator==2.1.1
openai==2.6.1
langchain==1.0.2
langchain-openai==1.0.1
langchain-community==0.4
tiktoken==0.12.0
faiss-cpu==1.12.0
langsmith>=0.3.45
requests>=2.32.5
beautifulsoup4==4.12.3
PyMuPDF==1.24.1
python-multipart==0.0.9
trafilatura>=1.7.0
playwright>=1.40.0

